+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 
 * Fix Electron support when `nodeIntegration` is disabled.
 
+### Dart API
+
+* All range checks for `SassColor` constructors now throw `RangeError`s with
+  `start` and `end` set.
+
 ## 1.32.5
 
 * **Potentially breaking bug fix:** When using `@for` with numbers that have